+++ tests/test_demands.py
@@ -104,17 +104,6 @@ class HttpServiceTests(PatchedSessionTests):
             method='GET', url='http://localhost/authed-endpoint',
             allow_redirects=True, auth=('foo', 'bar'))
 
-    def test_json_requests_have_formatted_data(self):
-        self.service.post(
-            '/data-endpoint', send_as_json=True,
-            data={'back': 'forth', 'forever': True, 'snowman-quote': '"☃"'})
-        data = self.request.call_args[1]['data']
-        self.assertIn('"snowman-quote": "\\"\\u2603\\""', data)
-        self.assertIn('"forever": true', data)
-        self.assertEqual(
-            self.request.call_args[1]['headers']['Content-Type'],
-            'application/json;charset=utf-8')
-
     @patch('demands.log')
     def test_logs_authentication_when_provided(self, mock_log):
         service = HTTPServiceClient(